# specs/001-cleaning-company-platform/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Nuqaa Asir Cleaning Booking & Operations Platform API
  version: "1.0.0"
  description: >
    REST API for the customer booking portal and admin operations dashboard.
    Two roles only: CUSTOMER and ADMIN. All endpoints are under /api/v1.
servers:
  - url: /api/v1
security:
  - bearerAuth: []
...
paths:
...
  /service-areas/{id}:
    patch:
      tags: [Catalog]
      summary: Update service area (Admin)
      parameters: [ { $ref: "#/components/parameters/IdParam" } ]
      responses: { "200": { description: Updated } }

  # ---------------------------------------------------------------- Pricing
...
  /subscriptions/{id}:
    patch:
      tags: [Subscriptions]
      summary: Update subscription config (Admin)
      parameters: [ { $ref: "#/components/parameters/IdParam" } ]
      responses: { "200": { description: Updated } }
  /subscriptions/{id}/pause:
    post:
      tags: [Subscriptions]
      summary: Pause a subscription (Admin)
      parameters: [ { $ref: "#/components/parameters/IdParam" } ]
      responses: { "200": { description: Paused } }
  /subscriptions/{id}/resume:
    post:
      tags: [Subscriptions]
      summary: Resume a subscription (Admin)
      parameters: [ { $ref: "#/components/parameters/IdParam" } ]
      responses: { "200": { description: Resumed } }
  /subscriptions/{id}/cancel:
    post:
      tags: [Subscriptions]
      summary: Cancel a subscription; preserves history (Admin)
      parameters: [ { $ref: "#/components/parameters/IdParam" } ]
      responses: { "200": { description: Cancelled } }
...
  /bookings/{id}/payments:
    get:
      tags: [Payments]
      summary: List payments for a booking (owner or Admin)
      parameters: [ { $ref: "#/components/parameters/IdParam" } ]
      responses: { "200": { description: OK } }
    post:
      tags: [Payments]
      summary: Manually record a payment (Admin)
      parameters: [ { $ref: "#/components/parameters/IdParam" } ]
      requestBody: { content: { application/json: { schema: { $ref: "#/components/schemas/PaymentInput" } } } }
      responses: { "201": { description: Recorded } }
  /bookings/{id}/invoice:
    get:
      tags: [Payments]
      summary: Get the non-fiscal receipt/invoice for a booking (owner or Admin)
      parameters: [ { $ref: "#/components/parameters/IdParam" } ]
      responses: { "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/Invoice" } } } } }

  # ---------------------------------------------------------------- Notifications
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    IdParam:
      in: path
      name: id
      required: true
      schema: { type: string, format: uuid }
...
  schemas:
...
    PaymentInput:
      type: object
      required: [method, amount]
      properties:
        method: { type: string, enum: [CASH, BANK_TRANSFER, POS, COMPLIMENTARY, OTHER] }
        amount: { type: integer }
        reference: { type: string, nullable: true }
        paidAt: { type: string, format: date-time, nullable: true }
    Invoice:
      type: object
      properties:
        id: { type: string, format: uuid }
        invoiceNumber: { type: string }
        issuedAt: { type: string, format: date-time }
        subtotal: { type: integer }
        discount: { type: integer }
        travelFee: { type: integer }
        tax: { type: integer }
        total: { type: integer }
        pdfUrl: { type: string, nullable: true }